rate-service-refactoring: move out remote repo
