+++ query.sql
@@ -1,0 +1,1 @@
+select "id", "time_stamp", "height", "signature", "proofs", "tx_type", "tx_version", "sender", "sender_public_key", "amount_asset", "price_asset", "fee", "sell_matcher_fee", "buy_matcher_fee", "o1_id", "o1_version", "o1_time_stamp", "o1_expiration", "o1_signature", "o1_sender", "o1_sender_public_key", "o1_type", "o1_matcher_fee", "o1_matcher_fee_asset_id", "o2_id", "o2_version", "o2_time_stamp", "o2_expiration", "o2_signature", "o2_sender", "o2_sender_public_key", "o2_type", "o2_matcher_fee", "o2_matcher_fee_asset_id", t.price * 10^(-8 - p_dec.decimals + a_dec.decimals) as "price", t.amount * 10^(-a_dec.decimals) as "amount", t.o1_price * 10^(-8 - p_dec.decimals + a_dec.decimals) as "o1_price", t.o1_amount * 10^(-a_dec.decimals) as "o1_amount", t.o2_price * 10^(-8 - p_dec.decimals + a_dec.decimals) as "o2_price", t.o2_amount * 10^(-a_dec.decimals) as "o2_amount" from (select "t"."id" as "id", "t"."time_stamp" as "time_stamp", "t"."height" as "height", "t"."signature" as "signature", "t"."proofs" as "proofs", "t"."tx_type" as "tx_type", "t"."tx_version" as "tx_version", "t"."sender" as "sender", "t"."sender_public_key" as "sender_public_key", "t"."amount_asset" as "amount_asset", "t"."price_asset" as "price_asset", t.price as "price", t.amount as "amount", t.fee * 10^(-8) as "fee", t.sell_matcher_fee * 10^(-8) as "sell_matcher_fee", t.buy_matcher_fee * 10^(-8) as "buy_matcher_fee", order1->>'id' as "o1_id", order1->>'version' as "o1_version", text_timestamp_cast(order1->>'timestamp') as "o1_time_stamp", text_timestamp_cast(order1->>'expiration') as "o1_expiration", order1->>'signature' as "o1_signature", order1->>'sender' as "o1_sender", order1->>'senderPublicKey' as "o1_sender_public_key", order1->>'orderType' as "o1_type", (order1->>'price')::double precision as "o1_price", (order1->>'amount')::double precision as "o1_amount", (order1->>'matcherFee')::double precision * 10^(-8) as "o1_matcher_fee", order1->>'matcherFeeAssetId' as "o1_matcher_fee_asset_id", order2->>'id' as "o2_id", order2->>'version' as "o2_version", text_timestamp_cast(order2->>'timestamp') as "o2_time_stamp", text_timestamp_cast(order2->>'expiration') as "o2_expiration", order2->>'signature' as "o2_signature", order2->>'sender' as "o2_sender", order2->>'senderPublicKey' as "o2_sender_public_key", order2->>'orderType' as "o2_type", (order2->>'price')::double precision as "o2_price", (order2->>'amount')::double precision as "o2_amount", (order2->>'matcherFee')::double precision * 10^(-8) as "o2_matcher_fee", order2->>'matcherFeeAssetId' as "o2_matcher_fee_asset_id" from "txs_7" as "t" where array[order1->>'sender', order2->>'sender'] @> '{3PEDGbdcSYAorLoV4oDVSJPTezGUPFrZj8f}' and "t"."sender" = '3PJaDyprvekvPXPuAtxrapacuDJopgJRaU3' order by "time_stamp" asc, "id" asc limit 11) as "t" inner join "asset_decimals" as "a_dec" on "t"."amount_asset" = "a_dec"."asset_id" inner join "asset_decimals" as "p_dec" on "t"."price_asset" = "p_dec"."asset_id" order by "time_stamp" asc, "id" asc
