+++ src/services/rates/RateEstimator.ts
@@ -1,25 +1,17 @@
-import { Task, of as taskOf } from "folktale/concurrency/task";
+import { Task } from "folktale/concurrency/task";
 import { Maybe, of as maybeOf } from 'folktale/maybe';
 
-import {
-  identity,
-  map,
-  chain,
-} from 'ramda';
+import { identity } from 'ramda';
 
 import { tap } from "../../utils/tap";
 import { AssetIdsPair, RateInfo } from "../../types";
 import { AppError, DbError } from "../../errorHandling";
-import { PgDriver } from "db/driver";
-import * as knex from 'knex'
 
-import makeSql from './sql';
-import { partitionByPreCount } from './repo';
+import { partitionByPreCount, AsyncGet } from './repo';
 import  RateCache, { RateCacheKey } from './repo/impl/RateCache';
 import RateInfoLookup from './repo/impl/RateInfoLookup'
 import { maybeIsNone } from "./util";
-
-const pg = knex({ client: 'pg' });
+import { RemoteRequestParams } from "./repo/impl/RemoteRateRepo";
 
 type ReqAndRes<TReq, TRes> = {
   req: TReq,
@@ -29,7 +21,7 @@ type ReqAndRes<TReq, TRes> = {
 export default class RateEstimator {
   constructor(
     private readonly cache: RateCache,
-    private readonly pgp: PgDriver
+    private readonly remoteGet: AsyncGet<RemoteRequestParams, RateInfo[], DbError>
   ) { }
 
   estimate(pairs: AssetIdsPair[], matcher: string, timestamp: Maybe<Date>): Task<AppError, ReqAndRes<AssetIdsPair, RateInfo>[]> {
@@ -49,24 +41,7 @@ export default class RateEstimator {
     
     const { preCount, toBeRequested } = partitionByPreCount(this.cache, pairs, getCacheKey)
 
-    const pairsSqlParams = chain(it => [it.amountAsset, it.priceAsset], toBeRequested)
-
-    const sql = pg.raw(makeSql(toBeRequested.length), [ timestamp.getOrElse(new Date()), matcher, ...pairsSqlParams ])
-
-    // console.log(sql.toString())
-
-    const dbTask: Task<DbError, any[]> = toBeRequested.length === 0 ? taskOf([]) : this.pgp.any(sql.toString());
-
-    return dbTask.map(
-      (result: any[]): RateInfo[] =>  map(
-        (it: any): RateInfo => {
-          return {
-            amountAsset: it.amount_asset_id,
-            priceAsset: it.price_asset_id,
-            current: it.weighted_average_price
-          }
-        }, result)
-    ).map(
+    return this.remoteGet.get({ pairs: toBeRequested, matcher, timestamp }).map(
       tap(cacheAll)
     ).map(
       data => new RateInfoLookup(data.concat(preCount))
